[engraving] replaced QList to std. Part 12

+++ src/engraving/libmscore/page.cpp
@@ -109,7 +109,7 @@ void Page::draw(mu::draw::Painter* painter) const
     // draw header/footer
     //
 
-    int n = no() + 1 + score()->pageNumberOffset();
+    size_t n = no() + 1 + score()->pageNumberOffset();
     painter->setPen(curColor());
 
     QString s1, s2, s3;
@@ -229,7 +229,7 @@ qreal Page::headerExtension() const
         return 0.0;
     }
 
-    int n = no() + 1 + score()->pageNumberOffset();
+    size_t n = no() + 1 + score()->pageNumberOffset();
 
     QString s1, s2, s3;
 
@@ -272,7 +272,7 @@ qreal Page::footerExtension() const
         return 0.0;
     }
 
-    int n = no() + 1 + score()->pageNumberOffset();
+    size_t n = no() + 1 + score()->pageNumberOffset();
 
     QString s1, s2, s3;
 
@@ -417,7 +417,7 @@ QString Page::replaceTextMacros(const QString& s) const
             // FALLTHROUGH
             case 'P': // on all pages
             {
-                int no = _no + 1 + score()->pageNumberOffset();
+                size_t no = _no + 1 + score()->pageNumberOffset();
                 if (no > 0) {
                     d += QString("%1").arg(no);
                 }
